remove unused UP051_1 and rename UP051_2
